Speed up packaging
The packaging step was very inefficient especially for metagenomes
that have a lot of MAGs.  This was because the process was
serialized and each one has to read through the input files over
and over.  This changes it so each file is open and read once
and the output is multiplexed to the various outputs files.
It also parallelizes the tar file generation.

For one test case the previous way was taking 6-7 hours and it nows
runs in a few minutes.